refactor with iso behavior
